+++ src/tools/IPAddressTool.java
@@ -37,7 +37,7 @@ public class IPAddressTool {
         }
         long ipAddress = 0;
         for (int i = 0; i < 4; i++) {
-            ipAddress += (long) (Integer.parseInt(quads[i]) % 256) * (long) Math.pow(256, (double) (4 - i));
+            ipAddress += (long) (Integer.parseInt(quads[i]) % 256) * (long) Math.pow(256, 4 - i);
         }
         return ipAddress;
     }
@@ -53,8 +53,8 @@ public class IPAddressTool {
         StringBuilder ipAddress = new StringBuilder();
 
         for (int i = 0; i < 4; i++) {
-            int quad = (int) (longIP / (long) Math.pow(256, (double) (4 - i)));
-            longIP -= (long) quad * (long) Math.pow(256, (double) (4 - i));
+            int quad = (int) (longIP / (long) Math.pow(256, 4 - i));
+            longIP -= (long) quad * (long) Math.pow(256, 4 - i);
             if (i > 0) {
                 ipAddress.append(".");
             }
